VSCode settings to use Deno extension
Since the `javascript` implementation is being tested with Deno this
enables using Deno for intellisense. The VSCode built-in TypeScript
support doesn't recognize the imports ending in `.ts`, so most of the
language features don't work because the imports are not resolved.

This also adds a VSCode list of recommended extensions to help people
with installing Deno and Go support.